mcp-servers/yandex-wordstat-mcp/pyproject.toml
[project]
name = "yandex-wordstat-mcp"
version = "0.1.0"
description = "MCP server for Yandex Wordstat API - keyword frequency data"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.8.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
